added skinny wall splitting
quad of a skinny wall (by given max ratio) can be replaced be multiple chunkier stacked quads. Skinnyness can be checked with check_if_too_skinny() and a version of get_triangles() to return multiple sub-quads exists an works. I thought skinny walls were causing issues with meshmixer but after implementing this, the issues did not go away. So the new version of get_triangles() is currently commented out and not used.