temp files now use file names instead of file obejcts
so far I treated tmp files as buffers that happen to be on a file i.e. I used file objects for this. That didn't work with MP (could not pickle) so now I use the file name (string) and leave it to the lowest level to write the file and close it (or throw an exception)