fixed issue with resampled numpy array not writeable
I use PIL to have it resample a numpy array and return it. However, PIL seems to lock the resampled array, so rather than returning a pointer I now return a copy.  in numpy 1.16.2 this could be circumvented by setting it to writeable manually but in 1.16.5 that throws an error